Crate admin1_reports

Features

Handle the following Resources with ease from the central hub ...

activities (list and watch)

channels (stop)

customer usage reports (get)

user usage report (get)

Everything else about the reports API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.activities().watch(...).doit()
let r = hub.channels().stop(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
admin1_reports = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate admin1_reports;
use admin1_reports::Channel;
use admin1_reports::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use admin1_reports::Reports;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Reports::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: Channel = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.activities().watch(&req, "userKey", "applicationName")
             .start_time("labore")
             .page_token("sea")
             .max_results(-90)
             .filters("dolores")
             .event_name("gubergren")
             .end_time("sadipscing")
             .customer_id("aliquyam")
             .actor_ip_address("ea")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

Activities
JSON template for a collection of activites.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list activities (response)

Activity
JSON template for the activity resource.

This type is not used in any activity, and only used as part of another schema.

ActivityActor
User doing the action.

This type is not used in any activity, and only used as part of another schema.

ActivityEvents
Activity events.

This type is not used in any activity, and only used as part of another schema.

ActivityEventsParameters
Parameter value pairs for various applications.

This type is not used in any activity, and only used as part of another schema.

ActivityId
Unique identifier for each activity record.

This type is not used in any activity, and only used as part of another schema.

ActivityListMethodBuilder
Retrieves a list of activities for a specific customer and application.

ActivityMethodsBuilder
A builder providing access to all methods supported on activity resources. It is not used directly, but through the Reports hub.

ActivityWatchMethodBuilder
Push changes to activities

Channel
An notification channel used to watch for resource changes.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

watch activities (request|response)

stop channels (request)

ChannelMethodsBuilder
A builder providing access to all methods supported on channel resources. It is not used directly, but through the Reports hub.

ChannelStopMethodBuilder
Stop watching resources through this channel

CustomerUsageReportGetMethodBuilder
Retrieves a report which is a collection of properties / statistics for a specific customer.

CustomerUsageReportMethodsBuilder
A builder providing access to all methods supported on customerUsageReport resources. It is not used directly, but through the Reports hub.

Reports
Central instance to access all Reports related resource activities

UsageReport
JSON template for a usage report.

This type is not used in any activity, and only used as part of another schema.

UsageReportEntity
Information about the type of the item.

This type is not used in any activity, and only used as part of another schema.

UsageReportParameters
Parameter value pairs for various applications.

This type is not used in any activity, and only used as part of another schema.

UsageReports
JSON template for a collection of usage reports.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get user usage report (response)

get customer usage reports (response)

UsageReportsWarnings
Warnings if any.

This type is not used in any activity, and only used as part of another schema.

UsageReportsWarningsData
Key-Value pairs to give detailed information on the warning.

This type is not used in any activity, and only used as part of another schema.

UserUsageReportGetMethodBuilder
Retrieves a report which is a collection of properties / statistics for a set of users.

UserUsageReportMethodsBuilder
A builder providing access to all methods supported on userUsageReport resources. It is not used directly, but through the Reports hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.